+++ src/nd_db/core.clj
@@ -3,7 +3,9 @@
             [nd-db
              [io :as ndio]
              [index :as ndix]
-             [util :as ndut]])
+             [util :as ndut]]
+            [nd-db.core :as nddb]
+            [clojure.string :as str])
   (:import [java.io File RandomAccessFile BufferedReader]))
 
 (defn- raw-db
@@ -25,10 +27,10 @@
     (if (.isFile ^File (io/file serialized-filepath))
       (ndio/parse-db params serialized-filepath)
       (let [[_ serialized-filename] (ndio/path->folder+filename serialized-filepath)]
-        (-> params
-            (assoc :serialized-filename serialized-filename)
-            raw-db
-            ndio/serialize-db)))))
+        (->> (assoc params :serialized-filename serialized-filename)
+             raw-db
+             ndio/serialize-db
+             (ndix/re-index (:log-limit params)))))))
 
 (defn db
   "Tries to read the specified pre-parsed database from filesystem.
@@ -46,13 +48,14 @@
   :id-name        - Convenience parameter - if you just want to supply the name of the ID in the text
                     based data to search for - creates a regex under the hood. Should be used with
                     the next parameter for optimal speed.
+  :id-path        - Use with nippy databases. Docs can be indexed directly by path vector
   :id-type        - The type of data to store as ID (key) in the index
   :source-type    - If the source-type is different from the ID type to store in the index
   :index-folder   - Folder to persist index in, defaults to system temp folder
   :index-persist? - Set to false to inhibit storing the index on disk, defaults to true. Will also
                     inhibit the use of previously persisted indices!
   :filename       - .ndnippy input filename (full path)
-  :index-path     - Use with .ndnippy file, docs can be index directly by path vector"
+  :log-limit      - Read the documents log until and including this index (for versioning)"
   [& _params]
   {:post [(ndut/db? %)]}
   (let [{:keys [index-persist?] :as params} (apply ndio/parse-params _params)]
@@ -127,8 +130,9 @@
      (lazy-docs-eager-idx db (lazy-seq @(:index db)))))
   ([a b]
    {:pre [(every? (some-fn map? (partial instance? BufferedReader)) [a b])]}
-   (let [[db] (filter map? [a b])
-         [reader] (filter (partial instance? BufferedReader) [a b])]
+   (let [both [a b]
+         [db] (filter map? both)
+         [reader] (filter (partial instance? BufferedReader) both)]
      (lazy-docs-lazy-idx ndio/str->
                          (io/file (:filename db))
                          reader))))
@@ -156,3 +160,33 @@
     (lazy-ids-lazy-idx ndio/str-> i)
 
     :else (throw (ex-info "Pass either db or index-reader!" {:param-type (type i)}))))
+
+(defn- emit-docs [db ^String doc-emission-str]
+  "Emit a serialized document(s) to a database.
+NOT thread safe, only use this from a single thread,
+meaning DON'T do parallel writes to database..!"
+  {:pre [(ndut/db? db)
+         (string? doc-emission-str)]}
+  (with-open [w (ndio/append-writer (:filename db))]
+    (.write w doc-emission-str)
+    (.newLine w)
+    (.flush w))
+  db)
+
+(defn append [{:keys [doc-emitter log-limit] :as db} doc-or-docs]
+  {:pre [((some-fn map? sequential?) doc-or-docs)]}
+  (when (or (nil? doc-emitter)
+            log-limit)
+    (throw (ex-info "Can't write to historical database (when log-limit is set)!" {:log-limit log-limit})))
+  (let [docs (if (map? doc-or-docs)
+               [doc-or-docs]
+               doc-or-docs)]
+    (loop [all-docs-part (partition-all 128 docs)
+           aggr-db db]
+      (if (empty? all-docs-part)
+        aggr-db
+        (let [docs-part-stringed (map doc-emitter (first all-docs-part))]
+          (recur (rest all-docs-part)
+                 (-> aggr-db
+                     (emit-docs (->> docs-part-stringed (str/join "\n")))
+                     (ndix/append docs docs-part-stringed))))))))